Style the header as desired
Make the header a flex container by setting the display
to flex.
Center items inside the header vertically and horizontally
by adjusting flex-content to space-between and align-items to center. Add a padding of 10 pixels.

Add display of flex and gap of 10 pixels to the ul selector
and also remove the unnecessary default left padding.