Style the navigation as desired
Format the HTML structure by adding two divs (.left-part and .right-part).
Put the content that is supposed to be on the left to .left-part
and the rest to .right-part.

Create a rule of:
-display to flex,
-justify-content to center,
-align-items to center,
-and gap to 10 px

And apply it to the following class selectors:
header, links, left-part, and right-part.

This aligns all flex items inside these containers
horizontally and also centered.

Inside the header class selector
set justify-content to space-between to create space
between the left-part and right-part containers.

Add a padding of 10 pixels to the header.