Center align items on the page
Replace the text align center from the previous iteration.
Even though this would work it is best to use flexbox
as this exercise is all about practicing flexbox.

Add a new div (.main) that will contain all items on the page
excluding the footer.

Change the layout of .main to flexbox, change the direction to a column,
horizontally align items to the center and
add a gap between items of 32 pixels.