Restructure index.html
Change the structure of the modal so that it has three children,
.icon, .content, and .buttons.
This is done so it can be easy to change the layout to Flexbox and
apply the desired styles.

Inside content group buttons to its container (.buttons),
this will make the structure clean making it easier to add styles to it.